requirements.txt
streamlit
pandas
plotly
openpyxl
python-docx
gspread
google-auth
requests